Formation (#92)
* fix: user score dto and handle not found response

* feat: get all quizzes (/quizzes instead of /quiz)

* feat: /quizzes for list of scores

* refractor: plurals

* feat: uses user email in header

* refractor: 404 Not Found instead of throwing exception

* Update FormationController.java

* feat: videos can be set as watched

+++ authentication/src/main/java/com/learnandphish/authentication/AuthenticationController.java
@@ -4,7 +4,6 @@ import com.learnandphish.authentication.jwt.JWTUtil;
 import com.learnandphish.authentication.user.ScanResult;
 import com.learnandphish.authentication.jwt.JwtRequest;
 import com.learnandphish.authentication.jwt.JwtResponse;
-import com.learnandphish.authentication.config.RestTemplateConfig;
 import com.learnandphish.authentication.jwt.JwtUserDetailsService;
 import com.learnandphish.authentication.user.*;
 import jakarta.annotation.security.RolesAllowed;
@@ -131,7 +130,11 @@ public class AuthenticationController {
         UserData user = userDataRepository.findByEmail(email)
                 .stream()
                 .findFirst()
-                .orElseThrow(() -> new UsernameNotFoundException("User not found"));
+                .orElse(null);
+
+        if (user == null) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("User not found");
+        }
 
         if (!passwordEncoder.matches(request.getCurrentPassword(), user.getPasswordHash())) {
             return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Current password is incorrect");
@@ -149,11 +152,15 @@ public class AuthenticationController {
     }
 
     @PostMapping("/reset-password")
-    public ResponseEntity<?> resetPassword(@RequestBody String email) {
+    public ResponseEntity<?> resetPassword(@RequestParam String email) {
         UserData user = userDataRepository.findByEmail(email)
                 .stream()
                 .findFirst()
-                .orElseThrow(() -> new UsernameNotFoundException("User not found"));
+                .orElse(null);
+
+        if (user == null) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("User not found");
+        }
 
         String password = generatePassword();
         user.setPasswordHash(passwordEncoder.encode(password));
@@ -187,25 +194,33 @@ public class AuthenticationController {
 
     @RolesAllowed({"USER", "ADMIN"})
     @GetMapping("/need-change-password")
-    public ResponseEntity<Boolean> needChangePassword() {
+    public ResponseEntity<?> needChangePassword() {
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
         String email = authentication.getName();
 
         UserData user = userDataRepository.findByEmail(email)
                 .stream()
                 .findFirst()
-                .orElseThrow(() -> new UsernameNotFoundException("User not found"));
+                .orElse(null);
+
+        if (user == null) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("User not found");
+        }
 
         return ResponseEntity.ok(user.getChangePassword());
     }
 
     @RolesAllowed("ADMIN")
     @PostMapping("/delete-user")
-    public ResponseEntity<?> deleteUser(@RequestBody String email) {
+    public ResponseEntity<?> deleteUser(@RequestParam String email) {
         UserData user = userDataRepository.findByEmail(email)
                 .stream()
                 .findFirst()
-                .orElseThrow(() -> new UsernameNotFoundException("User not found"));
+                .orElse(null);
+
+        if (user == null) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("User not found");
+        }
 
         userDataRepository.delete(user);
 
@@ -293,7 +308,7 @@ public class AuthenticationController {
 
     @RolesAllowed({"USER", "ADMIN"})
     @GetMapping("/get-user")
-    public ResponseEntity<UserDTO> getUser(@RequestHeader("Authorization") String token) {
+    public ResponseEntity<?> getUser(@RequestHeader("Authorization") String token) {
         String extractedToken = token.substring(7);
         String email = jwtUtil.extractUsername(extractedToken);
         UserData user = userDataRepository.findByEmail(email)
@@ -302,7 +317,7 @@ public class AuthenticationController {
             .orElse(null);
 
         if (user == null) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("User not found");
         }
 
         UserDTO userDTO = new UserDTO();
@@ -386,7 +401,7 @@ public class AuthenticationController {
      */
     private boolean isValidPassword(String password) {
         if (password.length() < 8) return false;
-        if (!password.chars().anyMatch(Character::isUpperCase)) return false; // At least one uppercase letter
+        if (password.chars().noneMatch(Character::isUpperCase)) return false; // At least one uppercase letter
         return password.chars().anyMatch(Character::isDigit);   // At least one digit
     }
